test server use  SELENIUM_TESTSERVER_HOST (localhost is default) replaced use of "string"
SELENIUM_PATH is now optional. You don't need to run server when using other WebDrivers than Remote. All Drivers like Firefox, Chrome etc. that are available in selenium bindings for python are working directly with browsers. To run HTMLUnit, HTMLUnitWithJS or Opera you need to use Remote WebDriver (in python) anyway and use capability that is introduced in this pull request. So all my changes are backward compatibilie.

Now selenium server is started optionaly when SELENIUM_PATH is set in settings. It may be used if someone want test locally in Opera or with use of HTMLUnit etc. But others may want to use Remote to connect to other selenium host (like we do) so it is optional.

We tested it with SELENIUM_PATH unset using remote selenium hub to run test for following browsers: IE7-IE9, Opera, Firefox, Chrome and to run locally with SELENIUM_DRIVER set to Firefox, Chrome. Also we tested with SELENIUM_PATH set to run with capability set to HTMLUNIT, HTMLUNITJS and OPERA.